requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-multiselectfield
django-cors-headers
django-filter
pillow
requests
torch
transformers
accelerate
sentencepiece
python-dotenv
groq